+++ src/main/java/fr/estecka/variantscit/reload/UnbakedModule.java
@@ -1,0 +1,14 @@
+package fr.estecka.variantscit.reload;
+
+import fr.estecka.variantscit.VariantLibrary;
+import fr.estecka.variantscit.modulebakers.IBakedModule;
+import fr.estecka.variantscit.modulebakers.IModuleBaker;
+
+public record UnbakedModule<T>(
+	IModuleBaker<T> baker,
+	T parameters
+){
+	public IBakedModule Bake(VariantLibrary library){
+		return baker.Bake(library, parameters);
+	}
+}